spiel--tokenize: refactor
Explicitly skip articles.
Don't accumulate input as we go.

This should still be able to handle adjectives, as the described objects will be
in a pool of candidates.
Though, only the last adjective is considered. We could improve this by
adding new objects to the "described" pool instead of resetting it at every adjective.
Compound verbs, e.g. "flip out" can be handled by the verb's actions.
Compound adjectives can be hyphenated.